Moved CompiledShaderEffect, EffectParam and RenderContextDefaultState to their own files
